changed type of all variables to 32-bit float
git-svn-id: https://svn.emii.org.au/repos/AODN-BoM/trunk@7171 80f082c3-aa3d-4a8a-bc4c-16289f4b763f

+++ convert.py
@@ -15,7 +15,8 @@ inc.defaultAttributes = inc.attributesFromFile(BOMattribFile)
 
 # csv formats
 i = np.int32
-f = np.float64
+# f = np.float64
+f = np.float32
 form1 = [('Time', 'S24'),
 	('Hs',f),
 	('Hrms',f),
